Task30

+++ src/main/java/student_jurij_mardusevic/lesson_2/Lesson2_Task30.java
@@ -9,9 +9,10 @@ public class Lesson2_Task30 {
         System.out.println("Number 1 = " + numberOne);
         System.out.println("Number 2 = " + numberTwo);
 
-        double sum = numberOne + numberTwo;
+        float sum = (float) (numberOne + numberTwo);
 
         System.out.println("Sum = " + sum);
     }
 
 }
+
